requirements.txt
Flask==3.1.2
paho-mqtt==2.1.0
PyYAML==6.0.3
moderngl==5.12.0
moderngl-window==3.1.1
PyGLM==2.8.2
pygame==2.6.1
